# api/openapi/task-service.yaml
openapi: 3.0.3
info:
  title: DeepFocus Task Service
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /v1/tasks:
    post:
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    get:
      summary: List tasks
      operationId: listTasks
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
  /v1/tasks/{taskId}:
    get:
      summary: Get a task
      operationId: getTask
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    put:
      summary: Update a task
      operationId: updateTask
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: deleted
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        interval_plan:
          $ref: '#/components/schemas/IntervalPlan'
        version:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        interval_plan:
          $ref: '#/components/schemas/IntervalPlan'
      required:
        - title
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        version:
          type: integer
          format: int64
    IntervalPlan:
      type: object
      properties:
        plan_id:
          type: string
          format: uuid
        cycles:
          type: integer
        work_duration_secs:
          type: integer
        break_duration_secs:
          type: integer
      required:
        - plan_id
        - cycles